time + paths to files fixed
